fix(Roles + Venues Controllers): shared display delete error code
