+++ app/controllers/roles_controller.rb
@@ -6,4 +6,8 @@ class RolesController < ApplicationController
     config.columns = [:name, :description]
     columns[:name].required = true
   end
+
+  # This must be included after active_scaffold config because it overrides
+  # some of the methods added by the controller
+  include DisplayDeleteErrors
 end
