+++ lib/display_delete_errors.rb
@@ -1,0 +1,11 @@
+module DisplayDeleteErrors
+  def do_destroy
+    destroy_find_record
+    self.successful = @record.destroy
+
+    if @record.errors.count > 0
+      flash[:warning] = (@record.errors.messages[:base] || []).join(' ')
+      self.successful = false
+    end
+  end
+end
